pyproject.toml
[project]
name = "insurance-agent"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "aiosqlite>=0.22.1",
    "fastapi>=0.135.1",
    "google-adk>=1.27.1",
    "greenlet>=3.3.2",
    "python-dotenv>=1.2.2",
    "uvicorn[standard]>=0.41.0",
]
